pyproject.toml
[project]
name = "nivra-ai-agent"
version = "0.1.0"
description = "AI Agent for Nivra AI Healthcare Assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.1",
    "langchain-groq>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.23",
    "langchain-huggingface>=0.1.0",
    "chromadb>=0.5.11",
    "groq>=0.4.0",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "huggingface-hub>=0.20.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pillow>=12.1.0",
]
